Move tableprinter to `tui` package
