+++ internal/tui/tableprinter/tableprinter.go
